requirements.txt
google-ads
google-cloud-bigquery
pandas
google-auth-oauthlib